net/octeon_ep: fix device start
[ upstream commit c892964faa605c7884b454b435c4fb663dff0c9e ]

Currently the device start routine clears the TX doorbell and
updates RX doorbell, which causes the device start operation
to fail after a device stop.

Patch corrects the handling of RX & TX doorbell and ensures
that any pending packets in the RX queue are drained and gets
refilled before starting the device.

Fixes: 2c82554c98bd ("net/octeontx_ep: add device start and stop")

Signed-off-by: Vamsi Attunuru <vattunuru@marvell.com>