doc: fix display of commands in cpfl guide
[ upstream commit 484451b27293f2745baf6793e5c4a3c3224f50f5 ]

The literal text of the commands to use were not displayed properly
because of a double ":" at the end of the preceding text, meaning that
the ".. code-block" marker got recognised as the literal text to
include. Change the double colon to single, and also indent the line
continuation in the multi-line flow-create command.

Fixes: 03f976012304 ("net/cpfl: adapt FXP to flow engine")

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Acked-by: Praveen Shetty <praveen.shetty@intel.com>