bus/cdx: fix release in probing for secondary process
[ upstream commit ad13df8e9379c5f17e6548eca1ae829f71795fde ]

The variable cdx_addr is a stack-allocated zero-initialized char array, but
it is never used anywhere except in an error handler to stand it as a
device bus address for a VFIO device release call. This is incorrect, as it
should have been actual device address, so fix the code to reflect that.

Fixes: 45ef232af515 ("bus/cdx: introduce AMD CDX bus")

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Acked-by: Nipun Gupta <nipun.gupta@amd.com>