net/mlx5/hws: fix flow rule hash capability
[ upstream commit 970309c7aa889ee06178c6476fe153e17096fc25 ]

Some cases are not supported for rule hash calculation. For example
when the matcher defined as FW matcher, or is the hash type is different
than CRC32. One case is when the distribute mode is not by hash, the
previous condition checked a wrong capability, while this commit fixes
it to the correct check.

Fixes: 7f5e6de53aae ("net/mlx5/hws: query flow rule hash")

Signed-off-by: Itamar Gozlan <igozlan@nvidia.com>
Acked-by: Bing Zhao <bingz@nvidia.com>